Rename django app to mwmbl
