+++ mwmbl/urls.py
@@ -17,7 +17,7 @@ Including another URLconf
 from django.contrib import admin
 from django.urls import path
 
-from app.api import api
+from mwmbl.api import api
 
 urlpatterns = [
     path('admin/', admin.site.urls),
